requirements.txt
# YouTube automation pipeline dependencies
# Pin major versions; update regularly via Dependabot or manual review

# AI / LLM
anthropic>=0.40.0,<1.0.0

# Google APIs
google-api-python-client>=2.120.0,<3.0.0
google-auth>=2.30.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0

# ElevenLabs TTS
elevenlabs>=1.2.0,<2.0.0

# Google Trends
pytrends>=4.9.2,<5.0.0

# HTTP
requests>=2.31.0,<3.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-cov>=5.0.0,<6.0.0
